Refactors loading of router
Rangling in the code a bit, shifting the routers to a more
standard structure. Each router now managing their own concerns a bit more.
