+++ locksmith/package.json
@@ -69,6 +69,7 @@
     "winston": "^3.1.0"
   },
   "devDependencies": {
+    "@types/cors": "^2.8.5",
     "nodemon": "1.18.11",
     "sqlite3": "^4.0.6",
     "supertest": "^4.0.2",
